Reorganize
clean up file organization
