+++ requirements.txt
@@ -1,3 +1,4 @@
+python==3.0
 numpy==1.13
 matplotlib==2.0
 scipy==1.0
